Fix some type conversion warnings

+++ src/network/network_content.cpp
@@ -254,8 +254,8 @@ void ClientNetworkContentSocketHandler::RequestContentList(ContentVector *cv, bo
 
 	while (cv->size() > offset) {
 		Packet *p = new Packet(send_md5sum ? PACKET_CONTENT_CLIENT_INFO_EXTID_MD5 : PACKET_CONTENT_CLIENT_INFO_EXTID, TCP_MTU);
-		const uint to_send = (uint)std::min<uint>(cv->size() - offset, max_per_packet);
-		p->Send_uint8(to_send);
+		const uint to_send = std::min<uint>(static_cast<uint>(cv->size() - offset), max_per_packet);
+		p->Send_uint8(static_cast<uint8>(to_send));
 
 		for (uint i = 0; i < to_send; i++) {
 			const ContentInfo *ci = (*cv)[offset + i];
